+++ packages/plugin-e2e/src/fixtures/grafanaVersion.ts
@@ -1,13 +1,12 @@
-import { PlaywrightTestArgs, TestFixture } from '@playwright/test';
+import { TestFixture } from '@playwright/test';
+import { PlaywrightArgs } from '../types';
 
-type GrafanaVersion = TestFixture<string, PlaywrightTestArgs>;
+type GrafanaVersion = TestFixture<string, PlaywrightArgs>;
 
-export const grafanaVersion: GrafanaVersion = async ({ page }, use) => {
-  let grafanaVersion = process.env.GRAFANA_VERSION ?? '';
-  if (!grafanaVersion) {
-    await page.goto('/');
-    grafanaVersion = await page.evaluate('window.grafanaBootData.settings.buildInfo.version');
-  }
+export const grafanaVersion: GrafanaVersion = async ({ bootData }, use) => {
+  // plugins may override version in CI via env var
+  const version = process.env.GRAFANA_VERSION || bootData.version || '';
 
-  await use(grafanaVersion.replace(/\-.*/, ''));
+  // strip version suffix (e.g., "11.0.0-pre" -> "11.0.0")
+  await use(version.replace(/\-.*/, ''));
 };