+++ packages/plugin-e2e/src/fixtures/namespace.ts
@@ -1,0 +1,9 @@
+import { TestFixture } from '@playwright/test';
+import { PlaywrightArgs } from '../types';
+
+type Namespace = TestFixture<string, PlaywrightArgs>;
+
+export const namespace: Namespace = async ({ bootData }, use) => {
+  // default to 'default' if namespace is not available
+  await use(bootData.namespace || 'default');
+};