+++ CHANGELOG.md
@@ -1,3 +1,107 @@
+# (Tue Feb 10 2026)
+
+#### 🐛 Bug Fix
+
+- `@grafana/plugin-e2e@3.2.2`
+  - Plugin E2E: Bump @grafana/e2e-selectors to 12.4.0-21636304584 [#2432](https://github.com/grafana/plugin-tools/pull/2432) ([@grafana-plugins-platform-bot[bot]](https://github.com/grafana-plugins-platform-bot[bot]))
+
+#### Authors: 1
+
+- [@grafana-plugins-platform-bot[bot]](https://github.com/grafana-plugins-platform-bot[bot])
+
+---
+
+# (Fri Feb 06 2026)
+
+#### 🐛 Bug Fix
+
+- `@grafana/react-detect@0.5.2`
+  - fix: reduce false positives [#2434](https://github.com/grafana/plugin-tools/pull/2434) ([@jackw](https://github.com/jackw))
+- `@grafana/create-plugin@6.8.4`
+  - chore(deps): update dependency @grafana/plugin-e2e to v3.2.1 [#2427](https://github.com/grafana/plugin-tools/pull/2427) ([@renovate-sh-app[bot]](https://github.com/renovate-sh-app[bot]))
+
+#### Authors: 2
+
+- [@renovate-sh-app[bot]](https://github.com/renovate-sh-app[bot])
+- Jack Westbrook ([@jackw](https://github.com/jackw))
+
+---
+
+# (Fri Jan 30 2026)
+
+#### 🐛 Bug Fix
+
+- `@grafana/plugin-e2e@3.2.1`
+  - Plugin E2E: Bump @grafana/e2e-selectors to 12.4.0-21513454430 [#2423](https://github.com/grafana/plugin-tools/pull/2423) ([@grafana-plugins-platform-bot[bot]](https://github.com/grafana-plugins-platform-bot[bot]))
+- `@grafana/create-plugin@6.8.3`
+  - fix(deps): update dependency glob to v13 [#2392](https://github.com/grafana/plugin-tools/pull/2392) ([@renovate-sh-app[bot]](https://github.com/renovate-sh-app[bot]))
+
+#### Authors: 2
+
+- [@grafana-plugins-platform-bot[bot]](https://github.com/grafana-plugins-platform-bot[bot])
+- [@renovate-sh-app[bot]](https://github.com/renovate-sh-app[bot])
+
+---
+
+# (Fri Jan 30 2026)
+
+#### 🐛 Bug Fix
+
+- `@grafana/create-plugin@6.8.2`
+  - chore(deps): update create-plugin template github actions [#2415](https://github.com/grafana/plugin-tools/pull/2415) ([@renovate-sh-app[bot]](https://github.com/renovate-sh-app[bot]))
+
+#### Authors: 1
+
+- [@renovate-sh-app[bot]](https://github.com/renovate-sh-app[bot])
+
+---
+
+# (Fri Jan 30 2026)
+
+#### 🐛 Bug Fix
+
+- `@grafana/create-plugin@6.8.1`
+  - chore(deps): update dependency @grafana/plugin-e2e to v3.2.0 [#2421](https://github.com/grafana/plugin-tools/pull/2421) ([@renovate-sh-app[bot]](https://github.com/renovate-sh-app[bot]))
+
+#### Authors: 1
+
+- [@renovate-sh-app[bot]](https://github.com/renovate-sh-app[bot])
+
+---
+
+# (Thu Jan 29 2026)
+
+#### 🚀 Enhancement
+
+- `@grafana/plugin-e2e@3.2.0`
+  - Plugin E2E: OpenFeature support [#2408](https://github.com/grafana/plugin-tools/pull/2408) ([@sunker](https://github.com/sunker))
+
+#### 🐛 Bug Fix
+
+- `@grafana/create-plugin@6.8.0`
+  - chore(deps): update grafana patch dependencies to v12.3.2 [#2420](https://github.com/grafana/plugin-tools/pull/2420) ([@renovate-sh-app[bot]](https://github.com/renovate-sh-app[bot]))
+  - fix(deps): update dependency which to v6 [#2393](https://github.com/grafana/plugin-tools/pull/2393) ([@renovate-sh-app[bot]](https://github.com/renovate-sh-app[bot]))
+
+#### Authors: 2
+
+- [@renovate-sh-app[bot]](https://github.com/renovate-sh-app[bot])
+- Erik Sundell ([@sunker](https://github.com/sunker))
+
+---
+
+# (Fri Jan 23 2026)
+
+#### 🐛 Bug Fix
+
+- `@grafana/react-detect@0.5.1`
+  - fix: silence warnings so json flag prints valid json [#2417](https://github.com/grafana/plugin-tools/pull/2417) ([@jackw](https://github.com/jackw))
+
+#### Authors: 1
+
+- Jack Westbrook ([@jackw](https://github.com/jackw))
+
+---
+
 # (Thu Jan 22 2026)
 
 #### 🐛 Bug Fix